move max entries to settings
