+++ packages/@n8n/api-types/src/frontend-settings.ts
@@ -260,7 +260,6 @@ export type FrontendModuleSettings = {
 	'chat-hub'?: {
 		enabled: boolean;
 		providers: Record<ChatHubLLMProvider, ChatProviderSettingsDto>;
-		memory: string;
 	};
 
 	/**
