+++ packages/cli/src/modules/chat-hub/chat-hub-execution.service.ts
@@ -44,7 +44,6 @@ import type { NonStreamingResponseMode, ChatTriggerResponseMode } from './chat-h
 import { ChatHubMessageRepository } from './chat-message.repository';
 import { ChatStreamService } from './chat-stream.service';
 import { createStructuredChunkAggregator } from './stream-capturer';
-import { parseMessage } from '@n8n/chat-hub';
 
 @Service()
 export class ChatHubExecutionService {
@@ -249,7 +248,12 @@ export class ChatHubExecutionService {
 
 				// Extract and save any memory facts from the AI response
 				if (message.status === 'success' && message.content) {
-					await this.extractAndSaveMemoryItems(user.id, sessionId, message.id, message.content);
+					await this.chatHubMemoryService.extractAndSaveMemoryItems(
+						user.id,
+						sessionId,
+						message.id,
+						message.content,
+					);
 				}
 
 				// End the stream for this message
@@ -659,40 +663,6 @@ export class ChatHubExecutionService {
 		};
 
 		return { adapter: adapter as unknown as Response, waitForPendingOperations };
-	}
-
-	/**
-	 * Parse AI message content for memory commands and save/update items
-	 */
-	private async extractAndSaveMemoryItems(
-		userId: string,
-		sessionId: string,
-		messageId: string,
-		content: string,
-	): Promise<void> {
-		const chunks = parseMessage({ type: 'ai', content });
-		for (const chunk of chunks) {
-			if (chunk.type === 'memory-create' && !chunk.isIncomplete && chunk.item) {
-				try {
-					await this.chatHubMemoryService.addMemoryItem(userId, sessionId, messageId, chunk.item);
-				} catch (error) {
-					this.logger.warn('Failed to save memory item', {
-						cause: error instanceof Error ? error.message : String(error),
-					});
-				}
-			} else if (chunk.type === 'memory-edit' && !chunk.isIncomplete && chunk.item) {
-				try {
-					if (chunk.index > 0) {
-						await this.chatHubMemoryService.deleteMemoryItemByIndex(userId, chunk.index - 1);
-					}
-					await this.chatHubMemoryService.addMemoryItem(userId, sessionId, messageId, chunk.item);
-				} catch (error) {
-					this.logger.warn('Failed to update memory item', {
-						cause: error instanceof Error ? error.message : String(error),
-					});
-				}
-			}
-		}
 	}
 
 	/**
